refactor(core): move filters to repository
